Final
I was unable to get two Errors cleared in the html. But i came across
this and it said theres nothing wrong with the html.

http://stackoverflow.com/questions/18549726/element-link-is-missing-requ
ired-attribute-property

I’m not sure if its 100% correct, but i asked a few people and they
couldn’t find anything wrong with in those lines of code.